+++ CHANGELOG.md
@@ -14,11 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Prettier config as last overwrite for all config packages
 - Create `@absolunet/eslint-config-pwastudio` - With PWA Studio rules
 - Create `@absolunet/eslint-config-spire` - With Spire rules
+- Recover `@absolunet/eslint-loader` in this repository
 
 ### Changed
 - Remove `core` rules treated by Prettier
 - Drop support for Node.js 10
-- Remove `@absolunet/eslint-loader` in favor of vanilla code
 - Maintenance updates
 
 #### Update to ESLint 7.31.0